+++ src/main/kotlin/launcher/SplashScreen.form
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="launcher.SplashScreen">
-  <grid id="27dc6" binding="panel1" default-binding="true" layout-manager="BorderLayout" hgap="0" vgap="0">
+  <grid id="27dc6" binding="panel" layout-manager="BorderLayout" hgap="0" vgap="0">
     <constraints>
       <xy x="20" y="20" width="500" height="400"/>
     </constraints>
@@ -9,40 +9,18 @@
       <color color="-16777216"/>
     </border>
     <children>
-      <grid id="2216d" layout-manager="GridLayoutManager" row-count="1" column-count="2" same-size-horizontally="false" same-size-vertically="false" hgap="-1" vgap="-1">
-        <margin top="0" left="0" bottom="0" right="0"/>
-        <constraints border-constraint="North"/>
-        <properties/>
-        <border type="none"/>
-        <children>
-          <component id="61af0" class="javax.swing.JLabel">
-            <constraints>
-              <grid row="0" column="0" row-span="1" col-span="1" vsize-policy="0" hsize-policy="0" anchor="8" fill="0" indent="0" use-parent-layout="false"/>
-            </constraints>
-            <properties>
-              <text value="Label"/>
-            </properties>
-          </component>
-          <component id="f9651" class="javax.swing.JLabel">
-            <constraints>
-              <grid row="0" column="1" row-span="1" col-span="1" vsize-policy="0" hsize-policy="0" anchor="8" fill="0" indent="0" use-parent-layout="false"/>
-            </constraints>
-            <properties>
-              <text value="Label"/>
-            </properties>
-          </component>
-        </children>
-      </grid>
       <grid id="5fafd" layout-manager="FlowLayout" hgap="5" vgap="5" flow-align="1">
         <constraints border-constraint="South"/>
         <properties/>
         <border type="none"/>
         <children>
-          <component id="d2f0b" class="javax.swing.JProgressBar" binding="progressBar1" default-binding="true">
+          <component id="d2f0b" class="javax.swing.JProgressBar" binding="progressBar">
             <constraints/>
-            <properties/>
+            <properties>
+              <minimum value="0"/>
+            </properties>
           </component>
-          <component id="6adf8" class="javax.swing.JLabel">
+          <component id="6adf8" class="javax.swing.JLabel" binding="progressText">
             <constraints/>
             <properties>
               <text value="Label"/>
@@ -50,6 +28,18 @@
           </component>
         </children>
       </grid>
+      <component id="553b8" class="javax.swing.JLabel" binding="progressTitle">
+        <constraints border-constraint="Center"/>
+        <properties>
+          <text value="Label"/>
+        </properties>
+      </component>
+      <component id="3da54" class="javax.swing.JLabel" binding="versionLabel">
+        <constraints border-constraint="North"/>
+        <properties>
+          <text value="Label"/>
+        </properties>
+      </component>
     </children>
   </grid>
 </form>
